pyproject.toml
[tool.poetry]
name = "family-budget"
version = "0.1.0"
description = ""
authors = ["zdpxdev <163684189+zdpxdev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.11"
djangorestframework = "^3.15.0"
django-model-utils = "^4.4.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
django-filter = "^24.1"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pytest-django = "^4.8.0"
factory-boy = "^3.3.0"
black = "^24.3.0"

[tool.black]
include = '\.pyi?$'
line_length = 88
target-version = ['py311']

[tool.isort]
include_trailing_comma = true
line_length = 88
multi_line_output = 3
profile = "black"
use_parentheses = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "family_budget.settings"
filterwarnings ="ignore::DeprecationWarning"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
